+++ src/ServiceProvider.php
@@ -3,9 +3,6 @@
 namespace Axn\Illuminate;
 
 use Illuminate\Support\ServiceProvider as BaseServiceProvider;
-use Collective\Html\FormBuilder;
-use Collective\Html\HtmlBuilder;
-use Illuminate\View\Compilers\BladeCompiler;
 
 class ServiceProvider extends BaseServiceProvider
 {
@@ -17,18 +14,12 @@ class ServiceProvider extends BaseServiceProvider
     public function boot()
     {
         // Blade Directives
-        $this->app->afterResolving('blade.compiler', function(BladeCompiler $blade) {
-            require __DIR__.'/blade-directives.php';
-        });
+        require __DIR__.'/blade-directives.php';
 
         // HTML Macros (LaravelCollective)
-        $this->app->afterResolving('html', function(HtmlBuilder $html) {
-            require __DIR__.'/html-macros.php';
-        });
+        require __DIR__.'/html-macros.php';
 
         // Form Macros (LaravelCollective)
-        $this->app->afterResolving('form', function(FormBuilder $form) {
-            require __DIR__.'/form-macros.php';
-        });
+        require __DIR__.'/form-macros.php';
     }
 }